// src/pages/markets/markets.enum.ts
export enum MarketSupplyLevel {
  SCARCE = 'SCARCE',
  LIMITED = 'LIMITED',
  MODERATE = 'MODERATE',
  HIGH = 'HIGH',
  ABUNDANT = 'ABUNDANT',
}

export enum MarketActivityLevel {
  WEAK = 'WEAK',
  GROWING = 'GROWING',
  STRONG = 'STRONG',
  RESTRICTED = 'RESTRICTED',
}

export enum TradeGoodType {
  EXPORT = 'EXPORT',
  IMPORT = 'IMPORT',
  EXCHANGE = 'EXCHANGE',
}